requirements.txt
asgiref==3.8.1
chardet==5.2.0
colorama==0.4.6
Django==5.1.1
pillow==10.4.0
pypng==0.20220715.0
qrcode==7.4.2
reportlab==4.2.2
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
